UPSTREAM: fs: limit filesystem stacking depth
Add a simple read-only counter to super_block that indicates how deep this
is in the stack of filesystems.  Previously ecryptfs was the only stackable
filesystem and it explicitly disallowed multiple layers of itself.

Overlayfs, however, can be stacked recursively and also may be stacked
on top of ecryptfs or vice versa.

To limit the kernel stack usage we must limit the depth of the
filesystem stack.  Initially the limit is set to 2.

References: CVE-2014-9922
Change-Id: I75135eccd6c4371a9e9fe7dfb134bb5a1dff8a78
Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>